Add createHistory
